Fixed rule recursion
