+++ .gitignore
@@ -21,7 +21,6 @@ rosetta-profiling/reports/
 **/xtend-gen/
 **/emf-gen/
 **/src/generated/
-**/xsemantics-gen/
 
 .antlr-generator-3.2.0-patch.jar
 com.regnosys.rosetta.web/
