+++ rosetta-lang/pom.xml
@@ -56,10 +56,6 @@
             <artifactId>org.eclipse.xtext</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.eclipse.xsemantics</groupId>
-            <artifactId>org.eclipse.xsemantics.runtime</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.eclipse.emf</groupId>
             <artifactId>org.eclipse.emf.ecore.xcore.lib</artifactId>
         </dependency>
@@ -121,12 +117,6 @@
                             </includes>
                         </fileset>
                         <fileset>
-                            <directory>${basedir}/../rosetta-lang/xsemantics-gen</directory>
-                            <includes>
-                                <include>**/*</include>
-                            </includes>
-                        </fileset>
-                        <fileset>
                             <directory>${basedir}/../rosetta-lang/xtend-gen</directory>
                             <includes>
                                 <include>**/*</include>
@@ -180,7 +170,7 @@
                 <artifactId>xtext-maven-plugin</artifactId>
                 <executions>
                     <execution>
-                        <id>xcore-and-xsemantics-generator</id>
+                        <id>xcore-generator</id>
                         <phase>generate-sources</phase>
                         <goals>
                             <goal>generate</goal>
@@ -190,7 +180,7 @@
                 <configuration>
                     <compilerSourceLevel>${maven.compiler.release}</compilerSourceLevel>
                     <compilerTargetLevel>${maven.compiler.release}</compilerTargetLevel>
-                    <!-- TODO: uncomment this line for faster builds once we upgrade Xtext. Also remove the `emf` and `xsemantics` folder from the clean plugin. -->
+                    <!-- TODO: uncomment this line for faster builds once we upgrade Xtext. Also remove the `emf` folder from the clean plugin. -->
                     <!-- <incrementalXtextBuild>true</incrementalXtextBuild> -->
                     <!--<classPathLookupFilter>^.*\.(emf|xtext).*$</classPathLookupFilter>-->
                     <addOutputDirectoriesToCompileSourceRoots>true</addOutputDirectoriesToCompileSourceRoots>
@@ -217,16 +207,6 @@
                                 <outputConfiguration>
                                     <outputDirectory>
                                         ${project.basedir}/emf-gen/main/java</outputDirectory>
-                                </outputConfiguration>
-                            </outputConfigurations>
-                        </language>
-                        <language>
-                            <setup>
-                                org.eclipse.xsemantics.dsl.XsemanticsStandaloneSetup</setup>
-                            <outputConfigurations>
-                                <outputConfiguration>
-                                    <outputDirectory>
-                                        ${project.basedir}/xsemantics-gen/main/java</outputDirectory>
                                 </outputConfiguration>
                             </outputConfigurations>
                         </language>
