+++ rosetta-lang/src/main/java/com/regnosys/rosetta/RosettaRuntimeModule.xtend
@@ -23,23 +23,14 @@ import org.eclipse.xtext.resource.impl.DefaultResourceDescriptionStrategy
 import org.eclipse.xtext.parser.IEncodingProvider
 import com.google.inject.Binder
 import org.eclipse.xtext.service.DispatchingProvider
-import com.regnosys.rosetta.utils.ImplicitVariableUtil
-import org.eclipse.xsemantics.runtime.validation.XsemanticsValidatorFilter
-import com.regnosys.rosetta.validation.RetainXsemanticsIssuesOnGeneratedInputsFilter
 import org.eclipse.xtext.conversion.IValueConverterService
 import com.regnosys.rosetta.parsing.RosettaValueConverterService
-import com.regnosys.rosetta.parsing.BigDecimalConverter
 import com.regnosys.rosetta.transgest.ModelLoader
 import com.regnosys.rosetta.transgest.ModelLoaderImpl
-import com.regnosys.rosetta.formatting2.RosettaExpressionFormatter
-import com.regnosys.rosetta.formatting2.FormattingUtil
 import javax.inject.Provider
-import com.regnosys.rosetta.generator.java.util.RecordJavaUtil
 import com.regnosys.rosetta.serialization.RosettaTransientValueService
 import org.eclipse.xtext.parsetree.reconstr.ITransientValueService
 import com.regnosys.rosetta.resource.RosettaResource
-import com.regnosys.rosetta.typing.RosettaTyping
-import com.regnosys.rosetta.typing.RosettaTypingAuxiliary
 import org.eclipse.xtext.validation.IResourceValidator
 import com.regnosys.rosetta.validation.CachingResourceValidator
 import com.regnosys.rosetta.config.RosettaConfiguration
@@ -51,15 +42,6 @@ import com.regnosys.rosetta.formatting2.ResourceFormatterService
 
 /* Use this class to register components to be used at runtime / without the Equinox extension registry.*/
 class RosettaRuntimeModule extends AbstractRosettaRuntimeModule {
-	
-	def void configureXsemanticsTypeSystem(Binder binder) {
-		// During a language server build, the following three classes are injected over and over again
-		// for each Rosetta resource. This means that code generation is spending up to 54% of its time
-		// just injecting these classes. By binding them as singletons, this time virtually disappears
-		// since they will only be instantiated once.
-		binder.bind(RosettaTyping).asEagerSingleton
-		binder.bind(RosettaTypingAuxiliary).asEagerSingleton
-	}
 	
 	override Class<? extends IFragmentProvider> bindIFragmentProvider() {
 		RosettaFragmentProvider
@@ -99,18 +81,8 @@ class RosettaRuntimeModule extends AbstractRosettaRuntimeModule {
         	.to(UTF8EncodingProvider);
     }
     
-    def Class<? extends ImplicitVariableUtil> bindImplicitVariableUtil() {
-    	ImplicitVariableUtil
-    }
-    def Class<? extends XsemanticsValidatorFilter> bindXsemanticsValidatorFilter() {
-    	RetainXsemanticsIssuesOnGeneratedInputsFilter
-    }
-    
     override Class<? extends IValueConverterService> bindIValueConverterService() {
     	RosettaValueConverterService
-    }
-    def Class<? extends BigDecimalConverter> bindBigDecimalConverter() {
-    	BigDecimalConverter
     }
 	
 	override Class<? extends XtextResource> bindXtextResource() {
@@ -122,19 +94,6 @@ class RosettaRuntimeModule extends AbstractRosettaRuntimeModule {
 	
 	def Class<? extends ModelLoader> bindModelLoader() {
 		ModelLoaderImpl
-	}
-	
-
-	def Class<? extends RosettaExpressionFormatter> bindRosettaExpressionFormatter() {
-		RosettaExpressionFormatter
-	}
-	
-	def Class<? extends FormattingUtil> bindFormattingUtil() {
-		FormattingUtil
-	}
-	
-	def Class<? extends RecordJavaUtil> bindRecordFeatureMap() {
-		RecordJavaUtil
 	}
 	
 	def Class<? extends IResourceValidator> bindIResourceValidator() {
