+++ rosetta-lang/src/main/java/com/regnosys/rosetta/types/RTypeFunction.java
@@ -30,20 +30,6 @@ public abstract class RTypeFunction extends AbstractModelSymbol {
 	public RTypeFunction(DottedPath namespace, String name) {
 		super(namespace, name);
 	}
-	
-	// TODO: limitation of Xsemantics, which doesn't support anonymous classes.
-	public static RTypeFunction create(DottedPath namespace, String name, Function<Map<String, RosettaValue>, RType> evaluate, Function<RType, Optional<LinkedHashMap<String, RosettaValue>>> reverse) {
-		return new RTypeFunction(namespace, name) {
-			@Override
-			public RType evaluate(Map<String, RosettaValue> arguments) {
-				return evaluate.apply(arguments);
-			}
-			@Override
-			public Optional<LinkedHashMap<String, RosettaValue>> reverse(RType type) {
-				return reverse.apply(type);
-			}
-		};
-	}
 
 	public abstract RType evaluate(Map<String, RosettaValue> arguments);
