--- rosetta-lang/src/main/java/com/regnosys/rosetta/types/RosettaAuxiliary.xsemantics
@@ -1,145 +1,0 @@
-system com.regnosys.rosetta.typing.RosettaTypingAuxiliary
-
-validatorExtends AbstractDeclarativeValidator
-
-import com.regnosys.rosetta.rosetta.RosettaCardinality
-import com.regnosys.rosetta.rosetta.RosettaEnumeration
-import com.regnosys.rosetta.rosetta.RosettaPackage
-import com.regnosys.rosetta.rosetta.simple.Data
-import com.regnosys.rosetta.rosetta.simple.SimplePackage
-import com.regnosys.rosetta.types.RDataType
-import com.regnosys.rosetta.types.REnumType
-import com.regnosys.rosetta.types.RType
-import com.regnosys.rosetta.types.TypeFactory
-import com.regnosys.rosetta.utils.ExpressionHelper
-import java.util.List
-
-import com.regnosys.rosetta.utils.ImplicitVariableUtil
-import com.regnosys.rosetta.types.builtin.RBuiltinTypeService
-import com.regnosys.rosetta.interpreter.RosettaInterpreter
-import com.regnosys.rosetta.types.RAliasType
-import com.regnosys.rosetta.utils.RosettaSimpleSystemSolver
-import java.util.function.BiFunction
-import org.eclipse.xtext.validation.AbstractDeclarativeValidator
-import org.eclipse.emf.ecore.resource.ResourceSet
-import com.regnosys.rosetta.rosetta.RosettaFeature
-import com.regnosys.rosetta.types.builtin.RRecordType
-import com.regnosys.rosetta.rosetta.RosettaRecordType
-import com.regnosys.rosetta.rosetta.RosettaEnumValue
-import com.regnosys.rosetta.utils.ModelIdProvider
-import com.regnosys.rosetta.types.TypeSystem
-import com.regnosys.rosetta.types.RChoiceType
-
-inject extension TypeFactory typeFactory
-inject extension ExpressionHelper exprHelper
-inject extension ImplicitVariableUtil implicitVarUtil
-inject extension RBuiltinTypeService builtinTypes
-inject extension ModelIdProvider modelIdProvider
-inject RosettaInterpreter interpreter
-inject RosettaSimpleSystemSolver systemSolver
-inject TypeSystem typeSystem
-
-
-auxiliary {
-	ancestors(Data t) : List<Data>
-	ancestorEnums(RosettaEnumeration t) : List<RosettaEnumeration>
-	overlap(RosettaCardinality c1, RosettaCardinality c2)
-	join(RType t1, RType t2) : RType
-	allEnumValues(RosettaEnumeration d) : Iterable<RosettaEnumValue>
-	mayBeEmpty(RDataType d)
-	
-	keepTypeAliasIfPossible(RType t1, RType t2, BiFunction<RType, RType, RType> combineUnderlyingTypes): RType
-	allFeatures(RType t, ResourceSet resourceSet) : Iterable<? extends RosettaFeature>
-}
-
-/****** AUXILIARY DEFINITIONS *******/
-auxiliary ancestors(Data t) {
-	getAll(t,
-		SimplePackage::eINSTANCE.data_SuperType,
-		SimplePackage::eINSTANCE.data_SuperType,
-		typeof(Data)
-	)
-}
-auxiliary ancestorEnums(RosettaEnumeration t) {
-	getAll(t,
-		RosettaPackage::eINSTANCE.rosettaEnumeration_Parent,
-		RosettaPackage::eINSTANCE.rosettaEnumeration_Parent,
-		typeof(RosettaEnumeration)
-	)
-}
-auxiliary overlap(RosettaCardinality c1, RosettaCardinality c2) {
-	(c1.unbounded || c1.sup >= c2.inf) && (c2.unbounded || c2.sup >= c1.inf)
-	or
-	return false
-}
-auxiliary join(RType t1, RType t2) {
-	return typeSystem.join(t1, t2)
-}
-auxiliary allEnumValues(RosettaEnumeration e) {
-	if (e.parent === null) {
-		return e.enumValues;
-	} else {
-		return allEnumValues(e.parent) + e.enumValues;
-	}
-}
-auxiliary mayBeEmpty(RDataType t) {
-	t.allAttributes.forall[
-		cardinality.minBound === 0
-	]
-}
-
-auxiliary keepTypeAliasIfPossible(RAliasType t1, RAliasType t2, BiFunction<RType, RType, RType> combineUnderlyingTypes) {
-	if (t1.typeFunction == t2.typeFunction) {
-		val typeFunc = t1.typeFunction
-		val underlier = keepTypeAliasIfPossible(t1.refersTo, t2.refersTo, combineUnderlyingTypes)
-		return typeFunc.reverse(underlier)
-			.<RType>map[args| new RAliasType(typeFunc, args, underlier)]
-			.orElse(underlier)
-	} else {
-		val superAliases = newArrayList
-		var RType curr = t1
-		while (curr instanceof RAliasType) {
-			superAliases.add(curr)
-			curr = curr.refersTo
-		}
-		curr = t2
-		while (curr instanceof RAliasType) {
-			val tf = curr.typeFunction
-			val match = superAliases.findFirst[tf == typeFunction]
-			if (match !== null) {
-				return keepTypeAliasIfPossible(match, curr, combineUnderlyingTypes)
-			}
-			curr = curr.refersTo
-		}
-		return keepTypeAliasIfPossible(t1.refersTo, t2.refersTo, combineUnderlyingTypes)
-	}
-}
-auxiliary keepTypeAliasIfPossible(RAliasType t1, RType t2, BiFunction<RType, RType, RType> combineUnderlyingTypes) {
-	keepTypeAliasIfPossible(t1.refersTo, t2, combineUnderlyingTypes)
-}
-auxiliary keepTypeAliasIfPossible(RType t1, RAliasType t2, BiFunction<RType, RType, RType> combineUnderlyingTypes) {
-	keepTypeAliasIfPossible(t1, t2.refersTo, combineUnderlyingTypes)
-}
-auxiliary keepTypeAliasIfPossible(RType t1, RType t2, BiFunction<RType, RType, RType> combineUnderlyingTypes) {
-	combineUnderlyingTypes.apply(t1, t2)
-}
-
-auxiliary allFeatures(RType t, ResourceSet resourceSet) {
-	switch t {
-		RDataType:
-			t.allNonOverridenAttributes.map[EObject]
-		RChoiceType:
-			t.asRDataType.allFeatures(resourceSet)
-		REnumType:
-			t.EObject.allEnumValues
-		RRecordType: {
-			if (resourceSet !== null) {
-				builtinTypes.toRosettaType(t, RosettaRecordType, resourceSet).features
-			} else {
-				#[]
-			}
-		}
-		default:
-			#[]
-	}
-}
